conf.d: Don't call `rbenv rehash` at initialisation
As of rbenv 1.0.0 `rbenv rehash` is automatically called after a
gem is installed or removed. This covers about 98% of the cases and
always calling `rbenv rehash` at shell initialisation is a needless
performance hit at this point.

If really necessary it's always possible to manually add this to
your shell's initialisation or just call `rbenv rehash` those 2%
of other cases if you ever run into them.