conf.d: Replicate `rbenv root` logic
Calling `rbenv root` causes the `rbenv` function to be called and
then go through the whole stack, doing all the things `rbenv` does
to eventually return us either the value of `$RBENV_ROOT` or
if that's empty `$HOME/.rbenv`.

By simply replicating this piece of code in Fish and not calling
`rbenv root` during the initialisation we manage to speed things
up even further.